probe: simplify API and use an LRU cache
Make the API simpler by changing register and unregister functions to
accept just a pid instead of a list. If a caller wants to add several
pids to the same handler, they can use a loop.

Get rid of the fragile golang map + reference counting and use an LRU
cache.