*: track file operations
This implements file operation tracking by hooking into fd_install and
reading `/proc/$PID/fd/$FD`.

To make sure we don't get incorrect results, we store the inode number
of the file (and the major and minor numbers of the device) and we check
they're the same in the file read from `/proc` in userspace. If they're
not, we just report "unknown" as the path. If the file doesn't exist
anymore, we report "[deleted]" and the possible path.

We store the `(pid, fd) -> path` mapping in a userspace Go map.

The idea for deletion is to delete the entry `(pid, fd)` when `pid`
exits.