bpf: switch to go-bindata
Instead of generating a BPF object file and using that in probe, we
store the contents in go code vie go-bindata so we don't depend on where
in the FS it is and tools that vendor this library can access it.

We should set up some kind of CI system to check that the go file is
updated when a developer updates the eBPF source.