tracer: don't process perf event in a goroutine
If we use a goroutine, we lose the ordering properties of the perf
events and we get overlapping prints of the events.

Let's not do that.

One drawback is that it reduces the number of events we can handle, but
we lose events anyway and we can revisit this later.