handle_tcp_close: do not ignore close events
Even if the connection is already closed, do not ignore a second close
event. It is useful to catch close events on localhost: connections can
be closed outside of the tcp_close code path there.

This will fix the issue discovered by the simple network tests PR:
https://github.com/ShiftLeftSecurity/traceleft/pull/93

It is visible with the nmap-nc version of netcat. In general, the 4
versions of netcat (GNU, BSD, nmap, busybox) have all different
behaviour so it is a good idea to test with each of them.