*: add a flags field in common event
And a first flag COMMON_EVENT_FLAG_INCOMPLETE_PROBE_READ. It signals
when a `bpf_probe_read()` reading a string failed, so we know the data
in the event will be incomplete.

This can happen if the string is near the border of a page with the next
page being unmapped. See
https://github.com/ShiftLeftSecurity/traceleft/issues/114#issuecomment-323970771
for an example.